Have compat-gettimeofday for windows
This makes it uniform with the other compat-* implementations,
similarly to the recent compat-getopt. Windows does not have this,
so say that in sysdep.h; for other systems, it's irrelevant.

While here, change the silly (struct timezone *)0 to plain NULL,
drop an unused function, and tweak some whitespace.
