+++ Makefile
@@ -64,6 +64,7 @@ HAVE_SRCS = \
 COMPAT_SRCS = \
 	compat-err.c		\
 	compat-getopt.c		\
+	compat-gettimeofday.c	\
 	compat-hsearch.c	\
 	compat-hsearch.h	\
 	compat-progname.c	\
@@ -72,6 +73,7 @@ COMPAT_SRCS = \
 COMPAT_OBJS = \
 	compat-err.o		\
 	compat-getopt.o		\
+	compat-gettimeofday.o	\
 	compat-hsearch.o	\
 	compat-progname.o	\
 	compat-strtonum.o
@@ -83,7 +85,6 @@ WINDOWS = \
 	win/rtptools.sln				\
 	win/rtpdump.vcxproj win/rtpplay.vcxproj		\
 	win/rtpsend.vcxproj win/rtptrans.vcxproj	\
-	win/gettimeofday.c win/gettimeofday.h		\
 	win/winsocklib.c
 
 DISTFILES = \
