+++ multimer.c
@@ -153,7 +153,7 @@ struct timeval *timer_set(struct timeval *interval,
 
   /* calculate expiration time */
   if (relative) {
-    (void) gettimeofday(&(tp->time), (struct timezone *)0);
+    (void) gettimeofday(&(tp->time), NULL);
     timeradd(&(tp->time), interval, &(tp->time));
     assert(tp->time.tv_usec < 1000000);
   }
@@ -199,7 +199,7 @@ struct timeval *timer_get(struct timeval *timeout)
     if (!timerQ) return (struct timeval *)0;
 
     /* check head of timer queue to see if timer has expired */
-    (void) gettimeofday(&now, (struct timezone *)0);
+    (void) gettimeofday(&now, NULL);
     if (timerless(&now, &timerQ->time)) { /* unexpired, calc timeout */
       timeout->tv_sec  = timerQ->time.tv_sec  - now.tv_sec;
       timeout->tv_usec = timerQ->time.tv_usec - now.tv_usec;
