+++ rtpsend.c
@@ -772,7 +772,7 @@ static int generate(char *text, char *data, double *time, int *type)
 static double gettimeofday_d(void)
 {
   struct timeval tv;
-  gettimeofday(&tv, (struct timezone *)0);
+  gettimeofday(&tv, NULL);
   return tv.tv_sec + tv.tv_usec / 1e6;
 } /* gettimeofday_d */
